Cambios

+++ index.php
@@ -19,6 +19,16 @@
 
     <?php
     // Mensaje de éxito
+
+    session_start(); // Inicia la sesión si aún no se ha iniciado
+
+    // Verificar si hay un mensaje de éxito almacenado en la variable de sesión
+    if (isset($_SESSION['mensaje_exito'])) {
+        echo "<p style='color: green; font-weight: bold;'>{$_SESSION['mensaje_exito']}</p>";
+        // Eliminar el mensaje de éxito después de mostrarlo para que no se muestre en futuras recargas de la página
+        unset($_SESSION['mensaje_exito']);
+    }
+
     $mensaje_exito = "";
 
     // Verificar si se envió el formulario para agregar un producto
@@ -34,19 +44,23 @@
         $sql = "INSERT INTO inventariot (producto, cantidad, unidad, fecha, factura, observaciones) VALUES ('$producto', $cantidad, '$unidad', '$fecha', '$factura', '$observaciones')";
 
         if ($conn->query($sql) === TRUE) {
-            // Actualizar el mensaje de éxito si la inserción fue exitosa
+          
             $mensaje_exito = "Producto agregado con éxito";
         } else {
-            // Si hubo un error en la inserción, mostrar mensaje de error
+            
             echo "Error: " . $sql . "<br>" . $conn->error;
         }
     }
 
     // Verificar si se envió el formulario para eliminar un producto
-    if(isset($_POST['eliminar_producto'])) {
-        $producto_id = $_POST['eliminar_producto_id'];
+if(isset($_POST['eliminar_producto'])) {
+    $producto_id = $_POST['eliminar_producto_id'];
+    
+    // Eliminar los registros relacionados en la tabla salidas
+    $sql_eliminar_salidas = "DELETE FROM salidas WHERE producto_id = $producto_id";
 
-        // Eliminar el producto de la tabla de inventario
+    if ($conn->query($sql_eliminar_salidas) === TRUE) {
+        // Ahora que los registros relacionados en la tabla salidas han sido eliminados, podemos proceder a eliminar el producto de la tabla inventariot
         $sql_eliminar = "DELETE FROM inventariot WHERE codigo = $producto_id";
 
         if ($conn->query($sql_eliminar) === TRUE) {
@@ -56,8 +70,11 @@
             // Si hubo un error en la eliminación, mostrar mensaje de error
             echo "Error al eliminar el producto: " . $conn->error;
         }
+    } else {
+        // Si hubo un error al eliminar los registros relacionados en la tabla salidas, mostrar mensaje de error
+        echo "Error al eliminar los registros relacionados en la tabla salidas: " . $conn->error;
     }
-
+}
     // Obtener datos de la tabla de inventario
     $sql_inventario = "SELECT * FROM inventariot";
     $result_inventario = $conn->query($sql_inventario);
